Update copy assets script

+++ scripts/copy_assets.sh
@@ -1,9 +1,22 @@
 SRC=$PROJECT_ASSETS_DIR/dinild-trimp
 DEST=./assets
+ARR="\x1b[32m>>>\x1b[0m"
+OK="\x1b[32m... OK\x1b[0m\n"
 
-set -x
-rm -rf ./assets
-mkdir ./assets
-cp -R $SRC/audio $DEST/audio
-cp -R $SRC/models $DEST/models
-cp -R $SRC/textures $DEST/textures
+copy_asset_dir () {
+  echo "$ARR Copy $1"
+  cp -LR $SRC/$1/ $DEST/$1
+  echo "$OK"
+}
+
+if [ $# -eq 0 ]
+  then
+    rm -rf ./assets
+    mkdir ./assets
+    copy_asset_dir audio
+    copy_asset_dir models
+    copy_asset_dir textures
+  else
+    rm -rf ./assets/$1
+    copy_asset_dir $1
+fi
